feat: EssayViewModel 분리
